requirements.txt
opencv-python
pygame
pywin32
mysql-connector-python
